Do not list contents of tmp/ directory if it does not exist.

+++ Configuration/scripts/exportToOverleaf.py
@@ -102,8 +102,9 @@ print "Document is of type " + documentType.name + " with number " + cadiNumber 
 print "Exporting document with requisite files...",
 sys.stdout.flush ()
 tmpDir = re.sub (r"Removing all contents of temporary directory: (.*)", r"\1", subprocess.check_output (["tdr", "clean"])).strip ().rstrip ()
-for f in os.listdir (tmpDir):
-    shutil.rmtree (tmpDir + "/" + f)
+if os.path.exists (tmpDir):
+    for f in os.listdir (tmpDir):
+        shutil.rmtree (tmpDir + "/" + f)
 tdrOutput = subprocess.check_output (["tdr", "--style=" + documentType.name.lower (), "--export", "b", cadiNumber])
 exportDir = None
 for line in tdrOutput.splitlines ():
